+++ src/modules/user/user.controller.ts
@@ -1,23 +1,31 @@
 import {
   Controller,
   Get,
-  Post,
   Body,
   Patch,
   Param,
   Delete,
+  UseInterceptors,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UserService } from './user.service';
-import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
+import { AuthUserInterceptor } from '../../common/interceptors/auth-user.interceptor';
+import { AuthUser } from '../../common/decorators/auth-user.decorator';
+import { users } from '@prisma/client';
+import { AuthUserId } from '../../common/decorators/auth-user-id.decorator';
 
+@UseGuards(JwtAuthGuard)
+@UseInterceptors(AuthUserInterceptor)
 @Controller('users')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    return this.userService.create(createUserDto);
+  @Get('me')
+  getMe(@AuthUser() user: users) {
+    return user;
   }
 
   @Get()
@@ -26,17 +34,24 @@ export class UserController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
+  findOne(@Param('id') id: string, @AuthUserId() userId: number) {
+    if (+id !== userId) throw new ForbiddenException('Access denied');
     return this.userService.findOne(+id);
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.userService.update(+id, updateUserDto);
+  update(
+    @Param('id') id: string,
+    @Body() dto: UpdateUserDto,
+    @AuthUserId() userId: number,
+  ) {
+    if (+id !== userId) throw new ForbiddenException('Access denied');
+    return this.userService.update(+id, dto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @AuthUserId() userId: number) {
+    if (+id !== userId) throw new ForbiddenException('Access denied');
     return this.userService.remove(+id);
   }
 }
